+++ apps/web/ui/folders/add-folder-form.tsx
@@ -101,7 +101,7 @@ export const AddFolderForm = ({ onSuccess, onCancel }: AddFolderFormProps) => {
             {step === 1 ? (
               <div className="mt-6">
                 <label>
-                  <span className="text-content-emphasis text-sm font-medium">
+                  <span className="text-content-emphasis block text-sm font-medium">
                     Name
                   </span>
                   <div className="mt-2 flex rounded-md">
@@ -126,8 +126,9 @@ export const AddFolderForm = ({ onSuccess, onCancel }: AddFolderFormProps) => {
                 </label>
 
                 <label className="mt-6 block">
-                  <span className="text-content-emphasis text-sm font-medium">
-                    Description
+                  <span className="text-content-emphasis block text-sm font-medium">
+                    Description{" "}
+                    <span className="text-content-subtle">(optional)</span>
                   </span>
                   <textarea
                     className="mt-2 block w-full rounded-md border-neutral-300 text-neutral-900 placeholder-neutral-400 focus:border-neutral-500 focus:outline-none focus:ring-neutral-500 sm:text-sm"
